Use csv module
By using the csv module, a lot of code was able to be removed, as well
as a dependency on curses. The unknown CSV format used before was the
unix dialect and support for it has been kept.

A few new dialects have
been added. To incorporate this change, a backwards compatible update to
the ABI has been provided.